live chart seems to be working - need to work on dispatcher efficiency

+++ RobotParameters/SensorData.cs
@@ -2,9 +2,9 @@
 {
     public class SensorData
     {
-        public int RawSpeed { get; set; }
-        public int RawCurrent { get; set; }
-        public int RawVoltage { get; set; }
-        public int RawTemperature { get; set; }
+        public short RawSpeed { get; set; }
+        public short RawCurrent { get; set; }
+        public short RawVoltage { get; set; }
+        public short RawTemperature { get; set; }
     }
 }
